modif bouton avec ajouts de props
