+++ packages/frontend/web/src/pages/main-menu.tsx
@@ -25,7 +25,6 @@ export default function MainMenu() {
           <div className='flex h-50 items-center justify-center'>
             <div className='flex flex-col items-center justify-center'>
               <Link to={'/create-single'}>
-                {' '}
                 <AddButton>{'+'}</AddButton>
               </Link>
               <h2 className='text-secondary pt-1 text-xl'>
